modifing scripts. Not done yet.

+++ bin/stop.sh
@@ -1,8 +1,23 @@
 #!/bin/sh
 
-pid=$(cat ./runtime/pids/rash_api.pid)
-echo "----------------------------------------------" &&
-echo "Killing rash_api process running on pid: $pid" &&
-kill "$pid" &&
-echo "Service has been stopped" &&
-echo "----------------------------------------------"
+COMMAND_LOGS=$PWD/runtime/logs/command.log
+curr_date=`date +"%D %T"`
+
+if [ -f $PWD/runtime/pids/rash_api.pid ]
+then
+    pid=$(cat ./runtime/pids/rash_api.pid)
+    echo "[$curr_date]: Started - [[Server Stop Script on $(hostname -f)]]" | tee -a $COMMAND_LOGS &&
+    echo "[$curr_date]:     Current Directory: $PWD" | tee -a $COMMAND_LOGS &&
+    echo "[$curr_date]:     Script: stop.sh" | tee -a $COMMAND_LOGS &&
+    echo "[$curr_date]:     Killing rash_api process running on pid: $pid" | tee -a $COMMAND_LOGS &&
+    kill "$pid" &&
+    echo "[$curr_date]: Completed - [[Server Start Script on $(hostname -f)]]" | tee -a $COMMAND_LOGS &&
+    echo "[$curr_date]: --------------------------------------------------------------------------" | tee -a $COMMAND_LOGS
+else
+    echo "[$curr_date]: Started - [[Server Stop Script on $(hostname -f)]]" | tee -a $COMMAND_LOGS &&
+    echo "[$curr_date]:     Current Directory: $PWD" | tee -a $COMMAND_LOGS &&
+    echo "[$curr_date]:     Script: stop.sh" | tee -a $COMMAND_LOGS &&
+    echo "[$curr_date]:         NOTICE: RASH_API IS NOT RUNNING. NO ACTION TAKEN." | tee -a $COMMAND_LOGS &&
+    echo "[$curr_date]: Completed - [[Server Start Script on $(hostname -f)]]" | tee -a $COMMAND_LOGS &&
+    echo "[$curr_date]: --------------------------------------------------------------------------" | tee -a $COMMAND_LOGS
+fi
